renamed api and api test file for usage statuses
